task 3 nefunctional

+++ headers/functii_liste.h
@@ -1,3 +1,3 @@
 /*#include "./tipuri_de_date.h"*/
-void addAtBeginning(Echipa **,char *,int ,int punctaj,Jucator *);
+void addAtBeginning(Echipa **,char *,int ,float punctaj,Jucator *);
 void eliminare_echipa (Echipa **,float min1);
